test(mcp): keep coverage green across the server.go split
Mirror the production file naming after refactor e0a58d5:

- server_test.go         — setupOfflineTest helpers + getSchema contract
- helpers_test.go        — filterSnap, injectMeta, metaJSONResult
- handlers_schema_test.go    — list/describe/search/find_related
- handlers_query_test.go     — validate/explain/check_migration/suggest
- handlers_lint_test.go      — lint_schema scopes and filters
- handlers_health_test.go    — compare_nodes/detect*/vacuum_health
- handlers_snapshot_test.go  — reload_schema candidate handling

Add tools_registration_test.go: round-trips the registered tool list
through ListTools and asserts every listed tool resolves to a handler.
Catches drift between tools.go (registration) and handlers_*.go
(implementations). Also pins the offline tool surface — online-only
tools (explain_query, refresh_schema, check_drift) must not appear
without a live pool.
